requirements.txt
dj-database-url==0.5.0
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
